requirements.txt
altair==4.2.0
config==0.5.1
pandas==1.4.4
Pillow==9.3.0
streamlit==1.15.0
streamlit_authenticator==0.1.5
textblob==0.17.1
tweepy==4.12.1
vaderSentiment==3.3.2
